+++ app/Providers.tsx
@@ -1,11 +1,10 @@
-'use client'
+"use client";
+import { SessionProvider } from "next-auth/react";
+import { NextUIProvider } from "@nextui-org/react";
 
-import {NextUIProvider} from '@nextui-org/react'
-
-export function NextProvider({children}: { children: React.ReactNode }) {
-  return (
-    <NextUIProvider>
-      {children}
-    </NextUIProvider>
-  )
+export function NextProvider({ children }: { children: React.ReactNode }) {
+  return <NextUIProvider>{children}</NextUIProvider>;
 }
+export const AuthProvider = ({ children }: { children: React.ReactNode }) => {
+  return <SessionProvider>{children}</SessionProvider>;
+};