+++ components/ui/NavItems.tsx
@@ -72,7 +72,7 @@ export const NavItems = ({
         ))}
         <button className="border text-sm font-medium relative border-neutral-200 dark:border-white/[0.2] text-black dark:text-white px-4 py-2 rounded-full">
           <span>
-            <Link href={"/login"}>Login</Link>
+            <Link href={"/api/auth/signin"}>Login</Link>
           </span>
           <span className="absolute inset-x-0 w-1/2 mx-auto -bottom-px  h-px"></span>
         </button>