+++ lib/redis.ts
@@ -1,0 +1,3 @@
+import { Redis } from "ioredis";
+
+export const redisClient = new Redis(process.env.REDIS_DB_URL as string);